+++ src/main/java/nextstep/AuthConfiguration.java
@@ -1,0 +1,33 @@
+package nextstep;
+
+import auth.JwtTokenProvider;
+import auth.LoginController;
+import auth.LoginService;
+import auth.UserDetailsDao;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AuthConfiguration {
+
+    private UserDetailsDao userDetailsDao;
+
+    public AuthConfiguration(UserDetailsDao userDetailsDao) {
+        this.userDetailsDao = userDetailsDao;
+    }
+
+    @Bean
+    public JwtTokenProvider jwtTokenProvider() {
+        return new JwtTokenProvider();
+    }
+
+    @Bean
+    public LoginService loginService(JwtTokenProvider jwtTokenProvider) {
+        return new LoginService(userDetailsDao, jwtTokenProvider);
+    }
+
+    @Bean
+    public LoginController loginController(LoginService loginService) {
+        return new LoginController(loginService);
+    }
+}
