+++ src/main/java/nextstep/WebMvcConfiguration.java
@@ -2,12 +2,9 @@ package nextstep;
 
 import auth.AdminInterceptor;
 import auth.JwtTokenProvider;
-import auth.LoginController;
 import auth.LoginMemberArgumentResolver;
 import auth.LoginService;
-import auth.UserDetailsDao;
 import java.util.List;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -15,34 +12,22 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
 
-    private final UserDetailsDao userDetailsDao;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public WebMvcConfiguration(UserDetailsDao userDetailsDao) {
-        this.userDetailsDao = userDetailsDao;
-    }
+    private final LoginService loginService;
 
-    @Bean
-    public JwtTokenProvider jwtTokenProvider() {
-        return new JwtTokenProvider();
-    }
-
-    @Bean
-    public LoginService loginService(JwtTokenProvider jwtTokenProvider) {
-        return new LoginService(userDetailsDao, jwtTokenProvider);
-    }
-
-    @Bean
-    public LoginController loginController(LoginService loginService) {
-        return new LoginController(loginService);
+    public WebMvcConfiguration(JwtTokenProvider jwtTokenProvider, LoginService loginService) {
+        this.jwtTokenProvider = jwtTokenProvider;
+        this.loginService = loginService;
     }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider())).addPathPatterns("/admin/**");
+        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider)).addPathPatterns("/admin/**");
     }
 
     @Override
     public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(new LoginMemberArgumentResolver(loginService(jwtTokenProvider())));
+        argumentResolvers.add(new LoginMemberArgumentResolver(loginService));
     }
 }
